feat: remove BaseClient classes in V2 generator
