+++ src/CodeGenerator.php
@@ -23,7 +23,6 @@ use Google\Generator\Collections\Vector;
 use Google\Generator\Generation\SnippetGenerator;
 use Google\Generator\Generation\BuildMethodFragmentGenerator;
 use Google\Generator\Generation\EmptyClientGenerator;
-use Google\Generator\Generation\EmptyClientV2Generator;
 use Google\Generator\Generation\EnumConstantGenerator;
 use Google\Generator\Generation\GapicMetadataGenerator;
 use Google\Generator\Generation\GapicClientGenerator;
@@ -331,14 +330,14 @@ class CodeGenerator
             if ($generateSnippets) {
                 $snippetFiles = SnippetGenerator::generate($licenseYear, $service);
 
-                $emptyClientName = $migrationMode == MigrationMode::MIGRATION_MODE_UNSPECIFIED || $migrationMode == MigrationMode::PRE_MIGRATION_SURFACE_ONLY ?
+                $clientName = $migrationMode == MigrationMode::MIGRATION_MODE_UNSPECIFIED || $migrationMode == MigrationMode::PRE_MIGRATION_SURFACE_ONLY ?
                     $service->emptyClientType->name :
-                    $service->emptyClientV2Type->name;
+                    $service->gapicClientV2Type->name;
 
                 foreach ($snippetFiles as $methodName => $snippetFile) {
                     $code = $snippetFile->toCode();
                     $code = Formatter::format($code, 100);
-                    yield ["samples/{$version}{$emptyClientName}/{$methodName}.php", $code];
+                    yield ["samples/{$version}{$clientName}/{$methodName}.php", $code];
                 }
             }
 
@@ -348,14 +347,7 @@ class CodeGenerator
                 $file = GapicClientV2Generator::generate($ctx, $service, $generateSnippets);
                 $code = $file->toCode();
                 $code = Formatter::format($code);
-                yield ["src/{$version}Client/BaseClient/{$service->gapicClientV2Type->name}.php", $code];
-
-                // Very thin service client wrapper, for manual code additions if required.
-                $ctx = new SourceFileContext($service->emptyClientV2Type->getNamespace(), $licenseYear);
-                $file = EmptyClientV2Generator::generate($ctx, $service);
-                $code = $file->toCode();
-                $code = Formatter::format($code);
-                yield ["src/{$version}Client/{$service->emptyClientV2Type->name}.php", $code];
+                yield ["src/{$version}Client/{$service->gapicClientV2Type->name}.php", $code];
 
                 // Unit tests.
                 $ctx = new SourceFileContext($service->unitTestsV2Type->getNamespace(), $licenseYear);
