--- tests/Integration/goldens/securitycenter/src/V1/Client/BaseClient/SecurityCenterBaseClient.php
@@ -1,1279 +1,0 @@
-<?php
-/*
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/*
- * GENERATED CODE WARNING
- * Generated by gapic-generator-php from the file
- * https://github.com/googleapis/googleapis/blob/master/google/cloud/securitycenter/v1/securitycenter_service.proto
- * Updates to the above are reflected here through a refresh process.
- */
-
-namespace Google\Cloud\SecurityCenter\V1\Client\BaseClient;
-
-use Google\ApiCore\ApiException;
-use Google\ApiCore\CredentialsWrapper;
-use Google\ApiCore\GapicClientTrait;
-use Google\ApiCore\LongRunning\OperationsClient;
-use Google\ApiCore\OperationResponse;
-use Google\ApiCore\PagedListResponse;
-use Google\ApiCore\ResourceHelperTrait;
-use Google\ApiCore\RetrySettings;
-use Google\ApiCore\Transport\TransportInterface;
-use Google\ApiCore\ValidationException;
-use Google\Auth\FetchAuthTokenInterface;
-use Google\Cloud\Iam\V1\GetIamPolicyRequest;
-use Google\Cloud\Iam\V1\Policy;
-use Google\Cloud\Iam\V1\SetIamPolicyRequest;
-use Google\Cloud\Iam\V1\TestIamPermissionsRequest;
-use Google\Cloud\Iam\V1\TestIamPermissionsResponse;
-use Google\Cloud\SecurityCenter\V1\CreateFindingRequest;
-use Google\Cloud\SecurityCenter\V1\CreateNotificationConfigRequest;
-use Google\Cloud\SecurityCenter\V1\CreateSourceRequest;
-use Google\Cloud\SecurityCenter\V1\DeleteNotificationConfigRequest;
-use Google\Cloud\SecurityCenter\V1\Finding;
-use Google\Cloud\SecurityCenter\V1\GetNotificationConfigRequest;
-use Google\Cloud\SecurityCenter\V1\GetOrganizationSettingsRequest;
-use Google\Cloud\SecurityCenter\V1\GetSourceRequest;
-use Google\Cloud\SecurityCenter\V1\GroupAssetsRequest;
-use Google\Cloud\SecurityCenter\V1\GroupFindingsRequest;
-use Google\Cloud\SecurityCenter\V1\ListAssetsRequest;
-use Google\Cloud\SecurityCenter\V1\ListFindingsRequest;
-use Google\Cloud\SecurityCenter\V1\ListNotificationConfigsRequest;
-use Google\Cloud\SecurityCenter\V1\ListSourcesRequest;
-use Google\Cloud\SecurityCenter\V1\NotificationConfig;
-use Google\Cloud\SecurityCenter\V1\OrganizationSettings;
-use Google\Cloud\SecurityCenter\V1\RunAssetDiscoveryRequest;
-use Google\Cloud\SecurityCenter\V1\SecurityMarks;
-use Google\Cloud\SecurityCenter\V1\SetFindingStateRequest;
-use Google\Cloud\SecurityCenter\V1\Source;
-use Google\Cloud\SecurityCenter\V1\UpdateFindingRequest;
-use Google\Cloud\SecurityCenter\V1\UpdateNotificationConfigRequest;
-use Google\Cloud\SecurityCenter\V1\UpdateOrganizationSettingsRequest;
-use Google\Cloud\SecurityCenter\V1\UpdateSecurityMarksRequest;
-use Google\Cloud\SecurityCenter\V1\UpdateSourceRequest;
-use Google\LongRunning\Operation;
-use GuzzleHttp\Promise\PromiseInterface;
-
-/**
- * Service Description: V1 APIs for Security Center service.
- *
- * This class provides the ability to make remote calls to the backing service through method
- * calls that map to API methods.
- *
- * Many parameters require resource names to be formatted in a particular way. To
- * assist with these names, this class includes a format method for each type of
- * name, and additionally a parseName method to extract the individual identifiers
- * contained within formatted names that are returned by the API.
- *
- * This class is currently experimental and may be subject to changes. See {@see
- * \Google\Cloud\SecurityCenter\V1\SecurityCenterClient} for the stable
- * implementation
- *
- * @experimental
- *
- * @internal
- *
- * @method PromiseInterface createFindingAsync(CreateFindingRequest $request, array $optionalArgs = [])
- * @method PromiseInterface createNotificationConfigAsync(CreateNotificationConfigRequest $request, array $optionalArgs = [])
- * @method PromiseInterface createSourceAsync(CreateSourceRequest $request, array $optionalArgs = [])
- * @method PromiseInterface deleteNotificationConfigAsync(DeleteNotificationConfigRequest $request, array $optionalArgs = [])
- * @method PromiseInterface getIamPolicyAsync(GetIamPolicyRequest $request, array $optionalArgs = [])
- * @method PromiseInterface getNotificationConfigAsync(GetNotificationConfigRequest $request, array $optionalArgs = [])
- * @method PromiseInterface getOrganizationSettingsAsync(GetOrganizationSettingsRequest $request, array $optionalArgs = [])
- * @method PromiseInterface getSourceAsync(GetSourceRequest $request, array $optionalArgs = [])
- * @method PromiseInterface groupAssetsAsync(GroupAssetsRequest $request, array $optionalArgs = [])
- * @method PromiseInterface groupFindingsAsync(GroupFindingsRequest $request, array $optionalArgs = [])
- * @method PromiseInterface listAssetsAsync(ListAssetsRequest $request, array $optionalArgs = [])
- * @method PromiseInterface listFindingsAsync(ListFindingsRequest $request, array $optionalArgs = [])
- * @method PromiseInterface listNotificationConfigsAsync(ListNotificationConfigsRequest $request, array $optionalArgs = [])
- * @method PromiseInterface listSourcesAsync(ListSourcesRequest $request, array $optionalArgs = [])
- * @method PromiseInterface runAssetDiscoveryAsync(RunAssetDiscoveryRequest $request, array $optionalArgs = [])
- * @method PromiseInterface setFindingStateAsync(SetFindingStateRequest $request, array $optionalArgs = [])
- * @method PromiseInterface setIamPolicyAsync(SetIamPolicyRequest $request, array $optionalArgs = [])
- * @method PromiseInterface testIamPermissionsAsync(TestIamPermissionsRequest $request, array $optionalArgs = [])
- * @method PromiseInterface updateFindingAsync(UpdateFindingRequest $request, array $optionalArgs = [])
- * @method PromiseInterface updateNotificationConfigAsync(UpdateNotificationConfigRequest $request, array $optionalArgs = [])
- * @method PromiseInterface updateOrganizationSettingsAsync(UpdateOrganizationSettingsRequest $request, array $optionalArgs = [])
- * @method PromiseInterface updateSecurityMarksAsync(UpdateSecurityMarksRequest $request, array $optionalArgs = [])
- * @method PromiseInterface updateSourceAsync(UpdateSourceRequest $request, array $optionalArgs = [])
- */
-abstract class SecurityCenterBaseClient
-{
-    use GapicClientTrait;
-    use ResourceHelperTrait;
-
-    /** The name of the service. */
-    private const SERVICE_NAME = 'google.cloud.securitycenter.v1.SecurityCenter';
-
-    /** The default address of the service. */
-    private const SERVICE_ADDRESS = 'securitycenter.googleapis.com';
-
-    /** The default port of the service. */
-    private const DEFAULT_SERVICE_PORT = 443;
-
-    /** The name of the code generator, to be included in the agent header. */
-    private const CODEGEN_NAME = 'gapic';
-
-    /** The default scopes required by the service. */
-    public static $serviceScopes = [
-        'https://www.googleapis.com/auth/cloud-platform',
-    ];
-
-    private $operationsClient;
-
-    private static function getClientDefaults()
-    {
-        return [
-            'serviceName' => self::SERVICE_NAME,
-            'apiEndpoint' => self::SERVICE_ADDRESS . ':' . self::DEFAULT_SERVICE_PORT,
-            'clientConfig' => __DIR__ . '/../../resources/security_center_client_config.json',
-            'descriptorsConfigPath' => __DIR__ . '/../../resources/security_center_descriptor_config.php',
-            'gcpApiConfigPath' => __DIR__ . '/../../resources/security_center_grpc_config.json',
-            'credentialsConfig' => [
-                'defaultScopes' => self::$serviceScopes,
-            ],
-            'transportConfig' => [
-                'rest' => [
-                    'restClientConfigPath' => __DIR__ . '/../../resources/security_center_rest_client_config.php',
-                ],
-            ],
-        ];
-    }
-
-    /**
-     * Return an OperationsClient object with the same endpoint as $this.
-     *
-     * @return OperationsClient
-     */
-    public function getOperationsClient()
-    {
-        return $this->operationsClient;
-    }
-
-    /**
-     * Resume an existing long running operation that was previously started by a long
-     * running API method. If $methodName is not provided, or does not match a long
-     * running API method, then the operation can still be resumed, but the
-     * OperationResponse object will not deserialize the final response.
-     *
-     * @param string $operationName The name of the long running operation
-     * @param string $methodName    The name of the method used to start the operation
-     *
-     * @return OperationResponse
-     */
-    public function resumeOperation($operationName, $methodName = null)
-    {
-        $options = isset($this->descriptors[$methodName]['longRunning']) ? $this->descriptors[$methodName]['longRunning'] : [];
-        $operation = new OperationResponse($operationName, $this->getOperationsClient(), $options);
-        $operation->reload();
-        return $operation;
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a finding
-     * resource.
-     *
-     * @param string $organization
-     * @param string $source
-     * @param string $finding
-     *
-     * @return string The formatted finding resource.
-     */
-    public static function findingName(string $organization, string $source, string $finding): string
-    {
-        return self::getPathTemplate('finding')->render([
-            'organization' => $organization,
-            'source' => $source,
-            'finding' => $finding,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a folder
-     * resource.
-     *
-     * @param string $folder
-     *
-     * @return string The formatted folder resource.
-     */
-    public static function folderName(string $folder): string
-    {
-        return self::getPathTemplate('folder')->render([
-            'folder' => $folder,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * folder_asset_securityMarks resource.
-     *
-     * @param string $folder
-     * @param string $asset
-     *
-     * @return string The formatted folder_asset_securityMarks resource.
-     */
-    public static function folderAssetSecurityMarksName(string $folder, string $asset): string
-    {
-        return self::getPathTemplate('folderAssetSecurityMarks')->render([
-            'folder' => $folder,
-            'asset' => $asset,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * folder_source resource.
-     *
-     * @param string $folder
-     * @param string $source
-     *
-     * @return string The formatted folder_source resource.
-     */
-    public static function folderSourceName(string $folder, string $source): string
-    {
-        return self::getPathTemplate('folderSource')->render([
-            'folder' => $folder,
-            'source' => $source,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * folder_source_finding resource.
-     *
-     * @param string $folder
-     * @param string $source
-     * @param string $finding
-     *
-     * @return string The formatted folder_source_finding resource.
-     */
-    public static function folderSourceFindingName(string $folder, string $source, string $finding): string
-    {
-        return self::getPathTemplate('folderSourceFinding')->render([
-            'folder' => $folder,
-            'source' => $source,
-            'finding' => $finding,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * folder_source_finding_securityMarks resource.
-     *
-     * @param string $folder
-     * @param string $source
-     * @param string $finding
-     *
-     * @return string The formatted folder_source_finding_securityMarks resource.
-     */
-    public static function folderSourceFindingSecurityMarksName(string $folder, string $source, string $finding): string
-    {
-        return self::getPathTemplate('folderSourceFindingSecurityMarks')->render([
-            'folder' => $folder,
-            'source' => $source,
-            'finding' => $finding,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * notification_config resource.
-     *
-     * @param string $organization
-     * @param string $notificationConfig
-     *
-     * @return string The formatted notification_config resource.
-     */
-    public static function notificationConfigName(string $organization, string $notificationConfig): string
-    {
-        return self::getPathTemplate('notificationConfig')->render([
-            'organization' => $organization,
-            'notification_config' => $notificationConfig,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a organization
-     * resource.
-     *
-     * @param string $organization
-     *
-     * @return string The formatted organization resource.
-     */
-    public static function organizationName(string $organization): string
-    {
-        return self::getPathTemplate('organization')->render([
-            'organization' => $organization,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * organization_asset_securityMarks resource.
-     *
-     * @param string $organization
-     * @param string $asset
-     *
-     * @return string The formatted organization_asset_securityMarks resource.
-     */
-    public static function organizationAssetSecurityMarksName(string $organization, string $asset): string
-    {
-        return self::getPathTemplate('organizationAssetSecurityMarks')->render([
-            'organization' => $organization,
-            'asset' => $asset,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * organization_settings resource.
-     *
-     * @param string $organization
-     *
-     * @return string The formatted organization_settings resource.
-     */
-    public static function organizationSettingsName(string $organization): string
-    {
-        return self::getPathTemplate('organizationSettings')->render([
-            'organization' => $organization,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * organization_source resource.
-     *
-     * @param string $organization
-     * @param string $source
-     *
-     * @return string The formatted organization_source resource.
-     */
-    public static function organizationSourceName(string $organization, string $source): string
-    {
-        return self::getPathTemplate('organizationSource')->render([
-            'organization' => $organization,
-            'source' => $source,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * organization_source_finding resource.
-     *
-     * @param string $organization
-     * @param string $source
-     * @param string $finding
-     *
-     * @return string The formatted organization_source_finding resource.
-     */
-    public static function organizationSourceFindingName(string $organization, string $source, string $finding): string
-    {
-        return self::getPathTemplate('organizationSourceFinding')->render([
-            'organization' => $organization,
-            'source' => $source,
-            'finding' => $finding,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * organization_source_finding_securityMarks resource.
-     *
-     * @param string $organization
-     * @param string $source
-     * @param string $finding
-     *
-     * @return string The formatted organization_source_finding_securityMarks resource.
-     */
-    public static function organizationSourceFindingSecurityMarksName(string $organization, string $source, string $finding): string
-    {
-        return self::getPathTemplate('organizationSourceFindingSecurityMarks')->render([
-            'organization' => $organization,
-            'source' => $source,
-            'finding' => $finding,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a project
-     * resource.
-     *
-     * @param string $project
-     *
-     * @return string The formatted project resource.
-     */
-    public static function projectName(string $project): string
-    {
-        return self::getPathTemplate('project')->render([
-            'project' => $project,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * project_asset_securityMarks resource.
-     *
-     * @param string $project
-     * @param string $asset
-     *
-     * @return string The formatted project_asset_securityMarks resource.
-     */
-    public static function projectAssetSecurityMarksName(string $project, string $asset): string
-    {
-        return self::getPathTemplate('projectAssetSecurityMarks')->render([
-            'project' => $project,
-            'asset' => $asset,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * project_source resource.
-     *
-     * @param string $project
-     * @param string $source
-     *
-     * @return string The formatted project_source resource.
-     */
-    public static function projectSourceName(string $project, string $source): string
-    {
-        return self::getPathTemplate('projectSource')->render([
-            'project' => $project,
-            'source' => $source,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * project_source_finding resource.
-     *
-     * @param string $project
-     * @param string $source
-     * @param string $finding
-     *
-     * @return string The formatted project_source_finding resource.
-     */
-    public static function projectSourceFindingName(string $project, string $source, string $finding): string
-    {
-        return self::getPathTemplate('projectSourceFinding')->render([
-            'project' => $project,
-            'source' => $source,
-            'finding' => $finding,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * project_source_finding_securityMarks resource.
-     *
-     * @param string $project
-     * @param string $source
-     * @param string $finding
-     *
-     * @return string The formatted project_source_finding_securityMarks resource.
-     */
-    public static function projectSourceFindingSecurityMarksName(string $project, string $source, string $finding): string
-    {
-        return self::getPathTemplate('projectSourceFindingSecurityMarks')->render([
-            'project' => $project,
-            'source' => $source,
-            'finding' => $finding,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a
-     * security_marks resource.
-     *
-     * @param string $organization
-     * @param string $asset
-     *
-     * @return string The formatted security_marks resource.
-     */
-    public static function securityMarksName(string $organization, string $asset): string
-    {
-        return self::getPathTemplate('securityMarks')->render([
-            'organization' => $organization,
-            'asset' => $asset,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a source
-     * resource.
-     *
-     * @param string $organization
-     * @param string $source
-     *
-     * @return string The formatted source resource.
-     */
-    public static function sourceName(string $organization, string $source): string
-    {
-        return self::getPathTemplate('source')->render([
-            'organization' => $organization,
-            'source' => $source,
-        ]);
-    }
-
-    /**
-     * Formats a string containing the fully-qualified path to represent a topic
-     * resource.
-     *
-     * @param string $project
-     * @param string $topic
-     *
-     * @return string The formatted topic resource.
-     */
-    public static function topicName(string $project, string $topic): string
-    {
-        return self::getPathTemplate('topic')->render([
-            'project' => $project,
-            'topic' => $topic,
-        ]);
-    }
-
-    /**
-     * Parses a formatted name string and returns an associative array of the components in the name.
-     * The following name formats are supported:
-     * Template: Pattern
-     * - finding: organizations/{organization}/sources/{source}/findings/{finding}
-     * - folder: folders/{folder}
-     * - folderAssetSecurityMarks: folders/{folder}/assets/{asset}/securityMarks
-     * - folderSource: folders/{folder}/sources/{source}
-     * - folderSourceFinding: folders/{folder}/sources/{source}/findings/{finding}
-     * - folderSourceFindingSecurityMarks: folders/{folder}/sources/{source}/findings/{finding}/securityMarks
-     * - notificationConfig: organizations/{organization}/notificationConfigs/{notification_config}
-     * - organization: organizations/{organization}
-     * - organizationAssetSecurityMarks: organizations/{organization}/assets/{asset}/securityMarks
-     * - organizationSettings: organizations/{organization}/organizationSettings
-     * - organizationSource: organizations/{organization}/sources/{source}
-     * - organizationSourceFinding: organizations/{organization}/sources/{source}/findings/{finding}
-     * - organizationSourceFindingSecurityMarks: organizations/{organization}/sources/{source}/findings/{finding}/securityMarks
-     * - project: projects/{project}
-     * - projectAssetSecurityMarks: projects/{project}/assets/{asset}/securityMarks
-     * - projectSource: projects/{project}/sources/{source}
-     * - projectSourceFinding: projects/{project}/sources/{source}/findings/{finding}
-     * - projectSourceFindingSecurityMarks: projects/{project}/sources/{source}/findings/{finding}/securityMarks
-     * - securityMarks: organizations/{organization}/assets/{asset}/securityMarks
-     * - source: organizations/{organization}/sources/{source}
-     * - topic: projects/{project}/topics/{topic}
-     *
-     * The optional $template argument can be supplied to specify a particular pattern,
-     * and must match one of the templates listed above. If no $template argument is
-     * provided, or if the $template argument does not match one of the templates
-     * listed, then parseName will check each of the supported templates, and return
-     * the first match.
-     *
-     * @param string $formattedName The formatted name string
-     * @param string $template      Optional name of template to match
-     *
-     * @return array An associative array from name component IDs to component values.
-     *
-     * @throws ValidationException If $formattedName could not be matched.
-     */
-    public static function parseName(string $formattedName, string $template = null): array
-    {
-        return self::parseFormattedName($formattedName, $template);
-    }
-
-    /**
-     * Constructor.
-     *
-     * @param array $options {
-     *     Optional. Options for configuring the service API wrapper.
-     *
-     *     @type string $apiEndpoint
-     *           The address of the API remote host. May optionally include the port, formatted
-     *           as "<uri>:<port>". Default 'securitycenter.googleapis.com:443'.
-     *     @type string|array|FetchAuthTokenInterface|CredentialsWrapper $credentials
-     *           The credentials to be used by the client to authorize API calls. This option
-     *           accepts either a path to a credentials file, or a decoded credentials file as a
-     *           PHP array.
-     *           *Advanced usage*: In addition, this option can also accept a pre-constructed
-     *           {@see \Google\Auth\FetchAuthTokenInterface} object or
-     *           {@see \Google\ApiCore\CredentialsWrapper} object. Note that when one of these
-     *           objects are provided, any settings in $credentialsConfig will be ignored.
-     *     @type array $credentialsConfig
-     *           Options used to configure credentials, including auth token caching, for the
-     *           client. For a full list of supporting configuration options, see
-     *           {@see \Google\ApiCore\CredentialsWrapper::build()} .
-     *     @type bool $disableRetries
-     *           Determines whether or not retries defined by the client configuration should be
-     *           disabled. Defaults to `false`.
-     *     @type string|array $clientConfig
-     *           Client method configuration, including retry settings. This option can be either
-     *           a path to a JSON file, or a PHP array containing the decoded JSON data. By
-     *           default this settings points to the default client config file, which is
-     *           provided in the resources folder.
-     *     @type string|TransportInterface $transport
-     *           The transport used for executing network requests. May be either the string
-     *           `rest` or `grpc`. Defaults to `grpc` if gRPC support is detected on the system.
-     *           *Advanced usage*: Additionally, it is possible to pass in an already
-     *           instantiated {@see \Google\ApiCore\Transport\TransportInterface} object. Note
-     *           that when this object is provided, any settings in $transportConfig, and any
-     *           $apiEndpoint setting, will be ignored.
-     *     @type array $transportConfig
-     *           Configuration options that will be used to construct the transport. Options for
-     *           each supported transport type should be passed in a key for that transport. For
-     *           example:
-     *           $transportConfig = [
-     *               'grpc' => [...],
-     *               'rest' => [...],
-     *           ];
-     *           See the {@see \Google\ApiCore\Transport\GrpcTransport::build()} and
-     *           {@see \Google\ApiCore\Transport\RestTransport::build()} methods for the
-     *           supported options.
-     *     @type callable $clientCertSource
-     *           A callable which returns the client cert as a string. This can be used to
-     *           provide a certificate and private key to the transport layer for mTLS.
-     * }
-     *
-     * @throws ValidationException
-     */
-    public function __construct(array $options = [])
-    {
-        $clientOptions = $this->buildClientOptions($options);
-        $this->setClientOptions($clientOptions);
-        $this->operationsClient = $this->createOperationsClient($clientOptions);
-    }
-
-    /** Handles execution of the async variants for each documented method. */
-    public function __call($method, $args)
-    {
-        if (substr($method, -5) !== 'Async') {
-            trigger_error('Call to undefined method ' . __CLASS__ . "::$method()", E_USER_ERROR);
-        }
-
-        array_unshift($args, substr($method, 0, -5));
-        return call_user_func_array([$this, 'startAsyncCall'], $args);
-    }
-
-    /**
-     * Creates a finding. The corresponding source must exist for finding creation
-     * to succeed.
-     *
-     * The async variant is {@see self::createFindingAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/create_finding.php
-     *
-     * @param CreateFindingRequest $request     A request to house fields associated with the call.
-     * @param array                $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return Finding
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function createFinding(CreateFindingRequest $request, array $callOptions = []): Finding
-    {
-        return $this->startApiCall('CreateFinding', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Creates a notification config.
-     *
-     * The async variant is {@see self::createNotificationConfigAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/create_notification_config.php
-     *
-     * @param CreateNotificationConfigRequest $request     A request to house fields associated with the call.
-     * @param array                           $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return NotificationConfig
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function createNotificationConfig(CreateNotificationConfigRequest $request, array $callOptions = []): NotificationConfig
-    {
-        return $this->startApiCall('CreateNotificationConfig', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Creates a source.
-     *
-     * The async variant is {@see self::createSourceAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/create_source.php
-     *
-     * @param CreateSourceRequest $request     A request to house fields associated with the call.
-     * @param array               $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return Source
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function createSource(CreateSourceRequest $request, array $callOptions = []): Source
-    {
-        return $this->startApiCall('CreateSource', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Deletes a notification config.
-     *
-     * The async variant is {@see self::deleteNotificationConfigAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/delete_notification_config.php
-     *
-     * @param DeleteNotificationConfigRequest $request     A request to house fields associated with the call.
-     * @param array                           $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function deleteNotificationConfig(DeleteNotificationConfigRequest $request, array $callOptions = []): void
-    {
-        $this->startApiCall('DeleteNotificationConfig', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Gets the access control policy on the specified Source.
-     *
-     * The async variant is {@see self::getIamPolicyAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/get_iam_policy.php
-     *
-     * @param GetIamPolicyRequest $request     A request to house fields associated with the call.
-     * @param array               $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return Policy
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function getIamPolicy(GetIamPolicyRequest $request, array $callOptions = []): Policy
-    {
-        return $this->startApiCall('GetIamPolicy', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Gets a notification config.
-     *
-     * The async variant is {@see self::getNotificationConfigAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/get_notification_config.php
-     *
-     * @param GetNotificationConfigRequest $request     A request to house fields associated with the call.
-     * @param array                        $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return NotificationConfig
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function getNotificationConfig(GetNotificationConfigRequest $request, array $callOptions = []): NotificationConfig
-    {
-        return $this->startApiCall('GetNotificationConfig', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Gets the settings for an organization.
-     *
-     * The async variant is {@see self::getOrganizationSettingsAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/get_organization_settings.php
-     *
-     * @param GetOrganizationSettingsRequest $request     A request to house fields associated with the call.
-     * @param array                          $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return OrganizationSettings
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function getOrganizationSettings(GetOrganizationSettingsRequest $request, array $callOptions = []): OrganizationSettings
-    {
-        return $this->startApiCall('GetOrganizationSettings', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Gets a source.
-     *
-     * The async variant is {@see self::getSourceAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/get_source.php
-     *
-     * @param GetSourceRequest $request     A request to house fields associated with the call.
-     * @param array            $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return Source
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function getSource(GetSourceRequest $request, array $callOptions = []): Source
-    {
-        return $this->startApiCall('GetSource', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Filters an organization's assets and  groups them by their specified
-     * properties.
-     *
-     * The async variant is {@see self::groupAssetsAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/group_assets.php
-     *
-     * @param GroupAssetsRequest $request     A request to house fields associated with the call.
-     * @param array              $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return PagedListResponse
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function groupAssets(GroupAssetsRequest $request, array $callOptions = []): PagedListResponse
-    {
-        return $this->startApiCall('GroupAssets', $request, $callOptions);
-    }
-
-    /**
-     * Filters an organization or source's findings and  groups them by their
-     * specified properties.
-     *
-     * To group across all sources provide a `-` as the source id.
-     * Example: /v1/organizations/{organization_id}/sources/-/findings,
-     * /v1/folders/{folder_id}/sources/-/findings,
-     * /v1/projects/{project_id}/sources/-/findings
-     *
-     * The async variant is {@see self::groupFindingsAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/group_findings.php
-     *
-     * @param GroupFindingsRequest $request     A request to house fields associated with the call.
-     * @param array                $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return PagedListResponse
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function groupFindings(GroupFindingsRequest $request, array $callOptions = []): PagedListResponse
-    {
-        return $this->startApiCall('GroupFindings', $request, $callOptions);
-    }
-
-    /**
-     * Lists an organization's assets.
-     *
-     * The async variant is {@see self::listAssetsAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/list_assets.php
-     *
-     * @param ListAssetsRequest $request     A request to house fields associated with the call.
-     * @param array             $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return PagedListResponse
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function listAssets(ListAssetsRequest $request, array $callOptions = []): PagedListResponse
-    {
-        return $this->startApiCall('ListAssets', $request, $callOptions);
-    }
-
-    /**
-     * Lists an organization or source's findings.
-     *
-     * To list across all sources provide a `-` as the source id.
-     * Example: /v1/organizations/{organization_id}/sources/-/findings
-     *
-     * The async variant is {@see self::listFindingsAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/list_findings.php
-     *
-     * @param ListFindingsRequest $request     A request to house fields associated with the call.
-     * @param array               $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return PagedListResponse
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function listFindings(ListFindingsRequest $request, array $callOptions = []): PagedListResponse
-    {
-        return $this->startApiCall('ListFindings', $request, $callOptions);
-    }
-
-    /**
-     * Lists notification configs.
-     *
-     * The async variant is {@see self::listNotificationConfigsAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/list_notification_configs.php
-     *
-     * @param ListNotificationConfigsRequest $request     A request to house fields associated with the call.
-     * @param array                          $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return PagedListResponse
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function listNotificationConfigs(ListNotificationConfigsRequest $request, array $callOptions = []): PagedListResponse
-    {
-        return $this->startApiCall('ListNotificationConfigs', $request, $callOptions);
-    }
-
-    /**
-     * Lists all sources belonging to an organization.
-     *
-     * The async variant is {@see self::listSourcesAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/list_sources.php
-     *
-     * @param ListSourcesRequest $request     A request to house fields associated with the call.
-     * @param array              $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return PagedListResponse
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function listSources(ListSourcesRequest $request, array $callOptions = []): PagedListResponse
-    {
-        return $this->startApiCall('ListSources', $request, $callOptions);
-    }
-
-    /**
-     * Runs asset discovery. The discovery is tracked with a long-running
-     * operation.
-     *
-     * This API can only be called with limited frequency for an organization. If
-     * it is called too frequently the caller will receive a TOO_MANY_REQUESTS
-     * error.
-     *
-     * The async variant is {@see self::runAssetDiscoveryAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/run_asset_discovery.php
-     *
-     * @param RunAssetDiscoveryRequest $request     A request to house fields associated with the call.
-     * @param array                    $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return OperationResponse
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function runAssetDiscovery(RunAssetDiscoveryRequest $request, array $callOptions = []): OperationResponse
-    {
-        return $this->startApiCall('RunAssetDiscovery', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Updates the state of a finding.
-     *
-     * The async variant is {@see self::setFindingStateAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/set_finding_state.php
-     *
-     * @param SetFindingStateRequest $request     A request to house fields associated with the call.
-     * @param array                  $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return Finding
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function setFindingState(SetFindingStateRequest $request, array $callOptions = []): Finding
-    {
-        return $this->startApiCall('SetFindingState', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Sets the access control policy on the specified Source.
-     *
-     * The async variant is {@see self::setIamPolicyAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/set_iam_policy.php
-     *
-     * @param SetIamPolicyRequest $request     A request to house fields associated with the call.
-     * @param array               $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return Policy
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function setIamPolicy(SetIamPolicyRequest $request, array $callOptions = []): Policy
-    {
-        return $this->startApiCall('SetIamPolicy', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Returns the permissions that a caller has on the specified source.
-     *
-     * The async variant is {@see self::testIamPermissionsAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/test_iam_permissions.php
-     *
-     * @param TestIamPermissionsRequest $request     A request to house fields associated with the call.
-     * @param array                     $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return TestIamPermissionsResponse
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function testIamPermissions(TestIamPermissionsRequest $request, array $callOptions = []): TestIamPermissionsResponse
-    {
-        return $this->startApiCall('TestIamPermissions', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Creates or updates a finding. The corresponding source must exist for a
-     * finding creation to succeed.
-     *
-     * The async variant is {@see self::updateFindingAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/update_finding.php
-     *
-     * @param UpdateFindingRequest $request     A request to house fields associated with the call.
-     * @param array                $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return Finding
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function updateFinding(UpdateFindingRequest $request, array $callOptions = []): Finding
-    {
-        return $this->startApiCall('UpdateFinding', $request, $callOptions)->wait();
-    }
-
-    /**
-     *
-     * Updates a notification config. The following update
-     * fields are allowed: description, pubsub_topic, streaming_config.filter
-     *
-     * The async variant is {@see self::updateNotificationConfigAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/update_notification_config.php
-     *
-     * @param UpdateNotificationConfigRequest $request     A request to house fields associated with the call.
-     * @param array                           $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return NotificationConfig
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function updateNotificationConfig(UpdateNotificationConfigRequest $request, array $callOptions = []): NotificationConfig
-    {
-        return $this->startApiCall('UpdateNotificationConfig', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Updates an organization's settings.
-     *
-     * The async variant is {@see self::updateOrganizationSettingsAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/update_organization_settings.php
-     *
-     * @param UpdateOrganizationSettingsRequest $request     A request to house fields associated with the call.
-     * @param array                             $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return OrganizationSettings
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function updateOrganizationSettings(UpdateOrganizationSettingsRequest $request, array $callOptions = []): OrganizationSettings
-    {
-        return $this->startApiCall('UpdateOrganizationSettings', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Updates security marks.
-     *
-     * The async variant is {@see self::updateSecurityMarksAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/update_security_marks.php
-     *
-     * @param UpdateSecurityMarksRequest $request     A request to house fields associated with the call.
-     * @param array                      $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return SecurityMarks
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function updateSecurityMarks(UpdateSecurityMarksRequest $request, array $callOptions = []): SecurityMarks
-    {
-        return $this->startApiCall('UpdateSecurityMarks', $request, $callOptions)->wait();
-    }
-
-    /**
-     * Updates a source.
-     *
-     * The async variant is {@see self::updateSourceAsync()} .
-     *
-     * @example samples/V1/SecurityCenterClient/update_source.php
-     *
-     * @param UpdateSourceRequest $request     A request to house fields associated with the call.
-     * @param array               $callOptions {
-     *     Optional.
-     *
-     *     @type RetrySettings|array $retrySettings
-     *           Retry settings to use for this call. Can be a {@see RetrySettings} object, or an
-     *           associative array of retry settings parameters. See the documentation on
-     *           {@see RetrySettings} for example usage.
-     * }
-     *
-     * @return Source
-     *
-     * @throws ApiException Thrown if the API call fails.
-     */
-    public function updateSource(UpdateSourceRequest $request, array $callOptions = []): Source
-    {
-        return $this->startApiCall('UpdateSource', $request, $callOptions)->wait();
-    }
-}
